+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to docs, or any other relevant information.
 # Changelog
 
 ## [Unreleased]
+- Add support for Workflow Updates as Nexus Operations 
 
 ### Added